fix: Tests working
