+++ dpypelines/pipeline/utils.py
@@ -49,6 +49,233 @@ def get_notifier():
         raise err
 
 
+def setup_clients():
+    """Set up clients for notification and email."""
+    notifier = get_notifier()
+    email_client = get_email_client()
+
+    if not notifier or not email_client:
+        err_msg = "Failed to set up notification or email client."
+        raise RuntimeError(err_msg)
+
+    logger.info(
+        "Clients set up successfully",
+        data={"notifier": notifier, "email_client": email_client},
+    )
+    return notifier, email_client
+
+
+def download_zip_file(s3_object_name: str) -> str:
+    """
+    Downloads a zip file from S3 into a local directory and returns the local file path.
+    """
+    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
+
+    # Split s3_object_name on final "/" in case of nested directory structure (e.g. <dir1>/dir2>/input.zip)
+    input_dir, input_zip_name = object_key.rsplit("/", maxsplit=1)
+    # Create a local directory to store downloaded zip file
+    Path(input_dir).mkdir(parents=True, exist_ok=True)
+
+    # Download S3 object to local directory
+    client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
+    with open(object_key, "wb") as f:
+        client.download_fileobj(Bucket=bucket_name, Key=object_key, Fileobj=f)
+    logger.info(
+        "Downloaded zip file to local folder",
+        data={"local_input_folder": input_dir, "local_file_name": input_zip_name},
+    )
+    return object_key
+
+
+def decompress_zip_file(local_zip_path: str) -> Path:
+    """
+    Decompress the given zip file into a local folder named after the input zip file (without extension).
+    """
+    # Create destination directory to store decompressed files
+    _, file_name = local_zip_path.rsplit("/", maxsplit=1)
+    destination_dir = Path(file_name.split(".")[0])
+    destination_dir.mkdir(parents=True, exist_ok=True)
+
+    # Extract zip file to destination directory
+    with zipfile.ZipFile(local_zip_path, "r") as f:
+        f.extractall(destination_dir)
+
+    # Handle different zip methods (zip files behave differently depending on operating system etc)
+    if any([f.is_dir() for f in destination_dir.rglob("*")]):
+        decompressed_file_dir = destination_dir / destination_dir
+    else:
+        decompressed_file_dir = destination_dir
+
+    logger.info(
+        "Decompressed zip file",
+        data={
+            "local_zip_path": str(local_zip_path),
+            "decompressed_file_dir": str(decompressed_file_dir),
+        },
+    )
+    return decompressed_file_dir
+
+
+def upload_to_s3_processing_folder(
+    s3_object_name: str,
+    local_object_key: str,
+    decompressed_file_dir: Path,
+) -> str:
+    """
+    Generate "processing" S3 folder name with timestamp and upload decompressed files/copy input zip file to the generated S3 folder.
+    """
+    bucket_name, s3_object_key = s3_object_name.split("/", maxsplit=1)
+    s3_processing_folder = f"processing/{datetime.now().strftime('%y-%m-%dT%H-%M')}-{decompressed_file_dir.parts[-1]}"
+    s3_client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
+
+    # Upload unzipped files to S3 "processing" folder
+    for file_path in decompressed_file_dir.rglob("*"):
+        # Delete hidden artefacts e.g. .DS_Store
+        if file_path.stem.startswith("."):
+            os.remove(file_path)
+        else:
+            s3_processing_object_name = (
+                f"{bucket_name}/{s3_processing_folder}/{file_path.name}"
+            )
+            upload_local_file_to_s3(
+                f"{decompressed_file_dir}/{file_path.name}",
+                s3_processing_object_name,
+                os.environ.get("AWS_PROFILE"),
+            )
+    logger.info("Decompressed files uploaded to S3 'processing' folder")
+
+    # Copy original zip file from S3 input location to S3 "processing" folder
+    s3_client.copy_object(
+        Bucket=bucket_name,
+        Key=f"{s3_processing_folder}/{local_object_key}",
+        CopySource={"Bucket": bucket_name, "Key": s3_object_key},
+    )
+
+    # Delete original zip file from S3 input location
+    s3_client.delete_object(Bucket=bucket_name, Key=s3_object_key)
+    logger.info(
+        "Input zip file copied to S3 'processing' folder and deleted from input folder"
+    )
+    return s3_processing_folder
+
+
+def copy_s3_processing_folder_to_processed_folder(
+    s3_object_name: str,
+    decompressed_file_dir: Path,
+    s3_processing_folder: str,
+) -> str:
+    """
+    Copy all files in S3 "processing" folder to S3 "processed" folder.
+    """
+    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
+    s3_processed_folder = f"processed/{datetime.now().strftime('%y-%m-%dT%H-%M')}-{decompressed_file_dir.parts[-1]}"
+    s3_client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
+
+    # Copy unzipped files from S3 "processing" folder to "processed" folder
+    for file_path in decompressed_file_dir.rglob("*"):
+        s3_client.copy_object(
+            Bucket=bucket_name,
+            Key=f"{s3_processed_folder}/{file_path.name}",
+            CopySource={
+                "Bucket": bucket_name,
+                "Key": f"{s3_processing_folder}/{file_path.name}",
+            },
+        )
+
+    # Copy original zip file from S3 "processing" folder to "processed" folder
+    s3_client.copy_object(
+        Bucket=bucket_name,
+        Key=f"{s3_processed_folder}/{object_key}",
+        CopySource={
+            "Bucket": bucket_name,
+            "Key": f"{s3_processing_folder}/{object_key}",
+        },
+    )
+    logger.info(
+        "Decompressed files and original zip submission copied to S3 'processed' folder",
+        data={"s3_processed_folder": s3_processed_folder},
+    )
+
+    return s3_processed_folder
+
+
+def delete_s3_processing_folder(
+    s3_object_name: str, decompressed_file_dir: Path, s3_processing_folder: str
+) -> None:
+    """
+    Delete all files from S3 "processing" folder to indicate successful submission.
+    """
+    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
+    s3_client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
+
+    # Delete unzipped files from S3 "processing" folder
+    for file_path in decompressed_file_dir.rglob("*"):
+        s3_client.delete_object(
+            Bucket=bucket_name, Key=f"{s3_processing_folder}/{file_path.name}"
+        )
+
+    # Delete zip file from S3 "processing" folder
+    s3_client.delete_object(
+        Bucket=bucket_name, Key=f"{s3_processing_folder}/{object_key}"
+    )
+    logger.info(
+        "Decompressed files and original zip submission deleted from S3 'processing' folder"
+    )
+
+
+def process_zip_file(s3_object_name: str) -> Tuple[LocalDirectoryStore, Path, str]:
+    """
+    Download a zip file from S3, decompress it, upload decompressed files to S3 'processing' folder and create a LocalDirectoryStore from the decompressed files.
+
+    Returns:
+        LocalDirectoryStore: A local store representing the processed folder.
+    """
+    # Download the zip file.
+    local_object_key = download_zip_file(s3_object_name)
+
+    # Decompress the zip file locally.
+    decompressed_file_dir = decompress_zip_file(local_object_key)
+
+    # Generate "processing" S3 folder name with timestamp and upload decompressed files
+    s3_processing_folder = upload_to_s3_processing_folder(
+        s3_object_name, local_object_key, decompressed_file_dir
+    )
+
+    # Validate that the decompressed folder contains files and create LocalDirectoryStore.
+    local_store = LocalDirectoryStore(decompressed_file_dir)
+    files = local_store.get_file_names()
+    if not files:
+        err_msg = f"Decompressed directory {decompressed_file_dir} is empty for s3_object_name {s3_object_name}. Available files: {files}"
+        raise FileNotFoundError(err_msg)
+
+    logger.info(
+        "S3 zip object processed successfully",
+        data={
+            "s3_object_name": s3_object_name,
+            "decompressed_file_dir": str(decompressed_file_dir),
+        },
+    )
+    return local_store, decompressed_file_dir, s3_processing_folder
+
+
+def validate_pipeline(files_dir: Path, pipeline_config: dict) -> dict:
+    """Validate the pipeline files against the configuration."""
+    validation_results = validate_pipeline_files(files_dir, pipeline_config)
+
+    try:
+        validation_results.get("manifest")
+    except Exception:
+        err_msg = f"Manifest validation failed for files in {files_dir} using config: {pipeline_config}."
+        raise ValidationException(
+            err_msg, files_dir=files_dir, validation_results=validation_results
+        )
+
+    logger.info(
+        "Pipeline validation completed successfully", data={"files_dir": str(files_dir)}
+    )
+    return validation_results
+
+
 def get_post_request_values_from_metadata(metadata: dict):
     """
     Generate the required path values and request body to submit to the Dataset API. This will be submitted as a POST request to the endpoint `/datasets/{dataset_path}/editions/{edition_path}/versions`
@@ -134,226 +361,6 @@ def get_distribution_details_for_request(dcat_distributions: list) -> list:
             "Error getting details of distributions for Dataset API request",
             distributions=dcat_distributions,
         ) from err
-
-
-def setup_clients():
-    """Set up clients for notification and email."""
-    notifier = get_notifier()
-    email_client = get_email_client()
-
-    if not notifier or not email_client:
-        err_msg = "Failed to set up notification or email client."
-        raise RuntimeError(err_msg)
-
-    logger.info(
-        "Clients set up successfully",
-        data={"notifier": notifier, "email_client": email_client},
-    )
-    return notifier, email_client
-
-
-def download_zip_file(s3_object_name: str) -> str:
-    """
-    Downloads a zip file from S3 into a local directory and returns the local file path.
-    """
-    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
-
-    # Split s3_object_name on final "/" in case of nested directory structure (e.g. <dir1>/dir2>/input.zip)
-    input_dir, input_zip_name = object_key.rsplit("/", maxsplit=1)
-    # Create a local directory to store downloaded zip file
-    Path(input_dir).mkdir(parents=True, exist_ok=True)
-
-    # Download S3 object to local directory
-    client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
-    with open(object_key, "wb") as f:
-        client.download_fileobj(Bucket=bucket_name, Key=object_key, Fileobj=f)
-    logger.info(
-        "Downloaded zip file to local folder",
-        data={"local_input_folder": input_dir, "local_file_name": input_zip_name},
-    )
-    return object_key
-
-
-def decompress_zip_file(local_zip_path: str) -> Path:
-    """
-    Decompress the given zip file into a local folder named after the input zip file (without extension).
-    """
-    # Create destination directory to store decompressed files
-    _, file_name = local_zip_path.rsplit("/", maxsplit=1)
-    destination_dir = Path(file_name.split(".")[0])
-    destination_dir.mkdir(parents=True, exist_ok=True)
-
-    # Extract zip file to destination directory
-    with zipfile.ZipFile(local_zip_path, "r") as f:
-        f.extractall(destination_dir)
-
-    # Handle different zip methods (zip files behave differently depending on operating system etc)
-    if any([f.is_dir() for f in destination_dir.rglob("*")]):
-        decompressed_file_dir = destination_dir / destination_dir
-    else:
-        decompressed_file_dir = destination_dir
-
-    logger.info(
-        "Decompressed zip file",
-        data={
-            "local_zip_path": str(local_zip_path),
-            "decompressed_file_dir": str(decompressed_file_dir),
-        },
-    )
-    return decompressed_file_dir
-
-
-def upload_to_s3_processing_folder(
-    s3_object_name: str,
-    local_object_key: str,
-    decompressed_file_dir: Path,
-) -> str:
-    """
-    Generate "processing" S3 folder name with timestamp and upload decompressed files/copy input zip file to the generated S3 folder.
-    """
-    bucket_name, s3_object_key = s3_object_name.split("/", maxsplit=1)
-    s3_processing_folder = f"processing/{datetime.now().strftime('%y-%m-%dT%H-%M')}-{decompressed_file_dir.parts[-1]}"
-    for file_path in decompressed_file_dir.rglob("*"):
-        # Delete hidden artefacts e.g. .DS_Store
-        if file_path.stem.startswith("."):
-            os.remove(file_path)
-        else:
-            s3_processing_object_name = (
-                f"{bucket_name}/{s3_processing_folder}/{file_path.name}"
-            )
-            upload_local_file_to_s3(
-                f"{decompressed_file_dir}/{file_path.name}",
-                s3_processing_object_name,
-                os.environ.get("AWS_PROFILE"),
-            )
-    logger.info("Decompressed files uploaded to S3 'processing' folder")
-
-    # Copy original zip file from S3 input location to S3 "processing" folder
-    client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
-    s3_destination_object_key = f"{s3_processing_folder}/{local_object_key}"
-    client.copy_object(
-        Bucket=bucket_name,
-        Key=s3_destination_object_key,
-        CopySource={"Bucket": bucket_name, "Key": s3_object_key},
-    )
-
-    # Delete original zip file from S3 input location
-    client.delete_object(Bucket=bucket_name, Key=s3_object_key)
-    logger.info(
-        "Input zip file copied to S3 'processing' folder and deleted from input folder"
-    )
-    return s3_processing_folder
-
-
-def copy_s3_processing_folder_to_processed_folder(
-    s3_object_name: str,
-    decompressed_file_dir: Path,
-    s3_processing_folder: str,
-) -> str:
-    """
-    Copy all files in S3 "processing" folder to S3 "processed" folder.
-    """
-    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
-    s3_client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
-    s3_processed_folder = f"processed/{datetime.now().strftime('%y-%m-%dT%H-%M')}-{decompressed_file_dir.parts[-1]}"
-
-    for file_path in decompressed_file_dir.rglob("*"):
-        s3_client.copy_object(
-            Bucket=bucket_name,
-            Key=f"{s3_processed_folder}/{file_path.name}",
-            CopySource={
-                "Bucket": bucket_name,
-                "Key": f"{s3_processing_folder}/{file_path.name}",
-            },
-        )
-
-    # Copy original zip file from S3 "processing" folder to S3 "processed" folder
-    s3_client.copy_object(
-        Bucket=bucket_name,
-        Key=f"{s3_processed_folder}/{object_key}",
-        CopySource={
-            "Bucket": bucket_name,
-            "Key": f"{s3_processing_folder}/{object_key}",
-        },
-    )
-    logger.info(
-        "Decompressed files and original zip submission copied to S3 'processed' folder",
-        data={"s3_processed_folder": s3_processed_folder},
-    )
-
-    return s3_processed_folder
-
-
-def delete_s3_processing_folder(
-    s3_object_name: str, decompressed_file_dir: Path, s3_processing_folder: str
-) -> None:
-    """
-    Delete all files from S3 "processing" folder to indicate successful submission.
-    """
-    bucket_name, object_key = s3_object_name.split("/", maxsplit=1)
-    client = _get_s3_client(profile_name=os.environ.get("AWS_PROFILE"))
-
-    for file_path in decompressed_file_dir.rglob("*"):
-        client.delete_object(
-            Bucket=bucket_name, Key=f"{s3_processing_folder}/{file_path.name}"
-        )
-    client.delete_object(Bucket=bucket_name, Key=f"{s3_processing_folder}/{object_key}")
-    logger.info(
-        "Decompressed files and original zip submission deleted from S3 'processing' folder"
-    )
-
-
-def process_zip_file(s3_object_name: str) -> Tuple[LocalDirectoryStore, Path, str]:
-    """
-    Download a zip file from S3, decompress it, upload decompressed files to S3 'processing' folder and create a LocalDirectoryStore from the decompressed files.
-
-    Returns:
-        LocalDirectoryStore: A local store representing the processed folder.
-    """
-    # Download the zip file.
-    local_object_key = download_zip_file(s3_object_name)
-
-    # Decompress the zip file locally.
-    decompressed_file_dir = decompress_zip_file(local_object_key)
-
-    # Generate "processing" S3 folder name with timestamp and upload decompressed files
-    s3_processing_folder = upload_to_s3_processing_folder(
-        s3_object_name, local_object_key, decompressed_file_dir
-    )
-
-    # Validate that the decompressed folder contains files and create LocalDirectoryStore.
-    local_store = LocalDirectoryStore(decompressed_file_dir)
-    files = local_store.get_file_names()
-    if not files:
-        err_msg = f"Decompressed directory {decompressed_file_dir} is empty for s3_object_name {s3_object_name}. Available files: {files}"
-        raise FileNotFoundError(err_msg)
-
-    logger.info(
-        "S3 zip object processed successfully",
-        data={
-            "s3_object_name": s3_object_name,
-            "decompressed_file_dir": str(decompressed_file_dir),
-        },
-    )
-    return local_store, decompressed_file_dir, s3_processing_folder
-
-
-def validate_pipeline(files_dir: Path, pipeline_config: dict) -> dict:
-    """Validate the pipeline files against the configuration."""
-    validation_results = validate_pipeline_files(files_dir, pipeline_config)
-
-    try:
-        validation_results.get("manifest")
-    except Exception:
-        err_msg = f"Manifest validation failed for files in {files_dir} using config: {pipeline_config}."
-        raise ValidationException(
-            err_msg, files_dir=files_dir, validation_results=validation_results
-        )
-
-    logger.info(
-        "Pipeline validation completed successfully", data={"files_dir": str(files_dir)}
-    )
-    return validation_results
 
 
 def upload_metadata(local_store, email_client, submitter_email):
